Added user to task relationship.

+++ TaskManagementApp/Controllers/TasksController.cs
@@ -2,6 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -13,16 +15,18 @@ namespace TaskManagementApp.Controllers
     public class TasksController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly UserManager<User> _userManager;
 
-        public TasksController(ApplicationDbContext context)
+        public TasksController(ApplicationDbContext context, UserManager<User> userManager)
         {
             _context = context;
+            _userManager = userManager;
         }
 
         // GET: Tasks
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Task.ToListAsync());
+            return View(await _context.Task.Include(t => t.User).ToListAsync());
         }
 
         // GET: Tasks/Details/5
@@ -33,8 +37,7 @@ namespace TaskManagementApp.Controllers
                 return NotFound();
             }
 
-            var task = await _context.Task
-                .FirstOrDefaultAsync(m => m.Id == id);
+            var task = await _context.Task.Include(t => t.User).FirstOrDefaultAsync(m => m.Id == id);
             if (task == null)
             {
                 return NotFound();
@@ -52,10 +55,11 @@ namespace TaskManagementApp.Controllers
         // POST: Tasks/ShowSearchResults
         public async Task<IActionResult> ShowSearchResults(String SearchPhrase)
         {
-            return View("Index", await _context.Task.Where( t => t.TaskDescription.Contains(SearchPhrase)).ToListAsync());
+            return View("Index", await _context.Task.Where(t => t.TaskDescription.Contains(SearchPhrase, StringComparison.CurrentCultureIgnoreCase)).ToListAsync());
         }
 
         // GET: Tasks/Create
+        [Authorize]
         public IActionResult Create()
         {
             return View();
@@ -64,10 +68,13 @@ namespace TaskManagementApp.Controllers
         // POST: Tasks/Create
         // To protect from overposting attacks, enable the specific properties you want to bind to, for 
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [Authorize]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,TaskDescription,Status")] Models.Task task)
         {
+            var user = _userManager.FindByNameAsync(User.Identity.Name).Result;
+            task.UserId = user.Id;
             if (ModelState.IsValid)
             {
                 _context.Add(task);
@@ -78,6 +85,7 @@ namespace TaskManagementApp.Controllers
         }
 
         // GET: Tasks/Edit/5
+        [Authorize]
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
@@ -86,6 +94,7 @@ namespace TaskManagementApp.Controllers
             }
 
             var task = await _context.Task.FindAsync(id);
+
             if (task == null)
             {
                 return NotFound();
@@ -96,6 +105,7 @@ namespace TaskManagementApp.Controllers
         // POST: Tasks/Edit/5
         // To protect from overposting attacks, enable the specific properties you want to bind to, for 
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [Authorize]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,TaskDescription,Status")] Models.Task task)
@@ -129,6 +139,7 @@ namespace TaskManagementApp.Controllers
         }
 
         // GET: Tasks/Delete/5
+        [Authorize]
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
@@ -138,6 +149,13 @@ namespace TaskManagementApp.Controllers
 
             var task = await _context.Task
                 .FirstOrDefaultAsync(m => m.Id == id);
+            var user = _userManager.FindByNameAsync(User.Identity.Name).Result;
+
+            if (task.UserId != user.Id)
+            {
+                return Unauthorized();
+            }
+
             if (task == null)
             {
                 return NotFound();
@@ -147,11 +165,19 @@ namespace TaskManagementApp.Controllers
         }
 
         // POST: Tasks/Delete/5
+        [Authorize]
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var task = await _context.Task.FindAsync(id);
+            var user = _userManager.FindByNameAsync(User.Identity.Name).Result;
+
+            if (task.UserId != user.Id)
+            {
+                return Unauthorized();
+            }
+
             _context.Task.Remove(task);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
